+++ src/Stitching/Stitching/Batching/IStitchingContext.cs
@@ -1,9 +1,11 @@
 ﻿using System;
+using System.Collections.Generic;
 using HotChocolate.Execution;
 
 namespace HotChocolate.Stitching
 {
     public interface IStitchingContext
+        : IObservable<IRemoteQueryClient>
     {
         IRemoteQueryClient GetRemoteQueryClient(string schemaName);
     }