+++ src/Stitching/Stitching/Introspection/Models/IntrospectionData.cs
@@ -1,0 +1,10 @@
+using Newtonsoft.Json;
+
+namespace HotChocolate.Stitching.Introspection.Models
+{
+    internal class IntrospectionData
+    {
+        [JsonProperty("__schema")]
+        public Schema Schema { get; set; }
+    }
+}